tests/wpt/tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.strokeRect.zero.1.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.strokeRect.zero.1</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.strokeRect.zero.1</h1>
<p class="desc">strokeRect of 0x0 pixels draws nothing</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/clear-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("strokeRect of 0x0 pixels draws nothing");
_addTest(function(canvas, ctx) {

  ctx.strokeStyle = '#f00';
  ctx.lineWidth = 250;
  ctx.strokeRect(50, 25, 0, 0);
  _assertPixel(canvas, 50,25, 0,0,0,0);

});
</script>
